refactor: move responses to internal
