+++ tests/auth_test.go
@@ -4,8 +4,8 @@ import (
 	"database/sql/driver"
 	"echo-demo-project/config"
 	"echo-demo-project/internal/models"
+	"echo-demo-project/internal/responses"
 	"echo-demo-project/requests"
-	"echo-demo-project/responses"
 	"echo-demo-project/server"
 	"echo-demo-project/server/handlers"
 	"echo-demo-project/services/token"
